Move metrics package to the right place
