+++ cmd/relayproxy/controller/all_flags.go
@@ -1,7 +1,7 @@
 package controller
 
 import (
-	"github.com/thomaspoignant/go-feature-flag/cmd/metric"
+	"github.com/thomaspoignant/go-feature-flag/cmd/relayproxy/metric"
 	"net/http"
 
 	"github.com/labstack/echo/v4"
